Refactor src files into two categories
Now that we are also extending core blocks, it made the most sense to
have them live in a folder beside our custom blocks. I've also removed
mojblocks.js as it was no longer being used.
