+++ src/custom-blocks/accordion/editor.scss
