+++ src/custom-blocks/accordion/style.scss
