+++ src/custom-blocks/banner/editor.scss
