+++ src/custom-blocks/banner/style.scss
