+++ src/custom-blocks/card/style.scss
