+++ src/custom-blocks/highlights-list/editor.scss
