+++ src/custom-blocks/highlights-list/style.scss
