+++ src/custom-blocks/index.js
@@ -1,0 +1,15 @@
+/**
+ * Import blocks as components.
+ */
+
+import "./highlights-list";
+import "./cta";
+import "./hero";
+import "./accordion";
+import "./staggered-box";
+import "./quote";
+import "./intro";
+import "./reveal";
+import "./card";
+import "./banner";
+
