+++ src/custom-blocks/quote/editor.scss
