+++ src/custom-blocks/quote/index.js
@@ -9,7 +9,7 @@ import { Fragment } from '@wordpress/element';
 import { RichText, MediaUpload, BlockControls, AlignmentToolbar } from '@wordpress/block-editor';
 import { Button, Dashicon } from '@wordpress/components';
 import domReady from '@wordpress/dom-ready';
-import Icons from '../../assets/svg/index';
+import Icons from '../../../assets/svg/index';
 
 const ALLOWED_MEDIA_TYPES = ['image'];
