+++ src/custom-blocks/quote/style.scss
