+++ src/custom-blocks/staggered-box/style.scss
