+++ src/extended-core-blocks/index.js
@@ -1,0 +1,4 @@
+/**
+ * Import extended block code
+ */
+ import "./file";
