+++ webpack.mix.js
@@ -22,8 +22,6 @@ mix_.webpackConfig({
 .sass('style-gutenburg.scss', 'style-gutenburg.css')
 .sass('assets/scss/ie.scss', 'ie.min.css');
 
-mix_.js('src/**/mojblocks.js', 'build/mojblocks.min.js');
-
 if (mix_.inProduction()) {
     mix_.version();
 } else {
